% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{all_sports}
\alias{all_sports}
\title{Data: hunting/fishing license history from customer sample}
\format{data frame}
\description{
Data: hunting/fishing license history from customer sample
}
\seealso{
Other Data: \code{\link{cust}}, \code{\link{lic}},
  \code{\link{retain_all}}, \code{\link{retain}},
  \code{\link{sale}}
}
\concept{Data}
\keyword{datasets}
