% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue.R
\name{compound_interest}
\alias{compound_interest}
\title{Calculate final principal given a rate of return}
\usage{
compound_interest(p0, r, t, n = NULL)
}
\arguments{
\item{p0}{original principal}

\item{r}{rate of return over some period (e.g., year)}

\item{t}{time the interest is applied (same units as r). Can be a scalar or
a vector}

\item{n}{compounding frequency. If NULL, will use continuous compounding}
}
\description{
https://en.wikipedia.org/wiki/Compound_interest
}
\examples{
# a wikipedia example
compound_interest(p0 = 1500, r = 0.043, t = 6, n = 4)
compound_interest(1500, 0.043, 6) # continuous
compound_interest(1500, 0.043, 0:6) # over time
}
\seealso{
Other estimating revenue: \code{\link{present_value_stream}},
  \code{\link{wsfr_lifetime_yrs}}, \code{\link{wsfr}}
}
\concept{estimating revenue}
