% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetime.R
\docType{data}
\name{cust}
\alias{cust}
\title{Data: 100K customer sample}
\format{data frame with 100K rows}
\description{
Data: 100K customer sample
}
\seealso{
Other Data: \code{\link{hunt}}, \code{\link{lic}},
  \code{\link{retain_all}}, \code{\link{retain}},
  \code{\link{sale}}
}
\concept{Data}
\keyword{datasets}
