% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetime.R
\docType{data}
\name{hunt}
\alias{hunt}
\title{Data: hunt history from 100K customer sample}
\format{data frame}
\description{
Data: hunt history from 100K customer sample
}
\seealso{
Other Data: \code{\link{cust}}, \code{\link{lic}},
  \code{\link{retain_all}}, \code{\link{retain}},
  \code{\link{sale}}
}
\concept{Data}
\keyword{datasets}
