% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{lic}
\alias{lic}
\title{Data: sportsman license types}
\format{data frame}
\description{
Data: sportsman license types
}
\seealso{
Other Data: \code{\link{all_sports}}, \code{\link{cust}},
  \code{\link{retain_all}}, \code{\link{retain}},
  \code{\link{sale}}
}
\concept{Data}
\keyword{datasets}
