% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-results.R
\name{nc_break_even}
\alias{nc_break_even}
\title{Find lifetime price (by current_age) where Revenue|A == Revenue|L}
\usage{
nc_break_even(retain_all, prices, return_life, inflation,
  wsfr_amount = NULL, min_amount = NULL, senior_price = NULL,
  senior_age = 65, perpetuity = TRUE, youth_ages = 0:15,
  age_cutoff = 80, max_price = 1000)
}
\arguments{
\item{retain_all}{data frame of predicted years by age like that produced by
\code{\link{nc_retain_all}}}

\item{prices}{data frame of lifetime prices by age with at least
2 variables: current_age and price_lifetime}

\item{return_life}{percentage return from lifetime fund}

\item{inflation}{inflation rate for depreciation of lifetime fund}

\item{wsfr_amount}{numeric estimated amount for aid dollars (use SFRF for
fishing, WRF for hunting)}

\item{min_amount}{numeric the minimum expenditure that will count for a
certified hunter/angler (use 2 for hunt/fish and 4 for combo)}

\item{senior_price}{numeric price for a senior lifetime license}

\item{senior_age}{numeric age when a participant will be expected to buy
a cheap lifetime license}

\item{perpetuity}{if TRUE, use a perpetuity calculation
(\url{https://en.wikipedia.org/wiki/Perpetuity}) instead of
\code{\link{present_value}}}

\item{youth_ages}{if not NULL, assumes for youths that the fund is able to
compound until adulthood (when the agency will begin drawing revenue). See
\code{\link{nc_price_lifetime_youth}} for details.}

\item{age_cutoff}{numeric final age that can be counted for lifetime
license-based WSFR dollars}

\item{max_price}{Maximum lifetime license price used in solver function. A
lower number will allow the function to execute more quickly.}
}
\description{
This uses a solver function, \code{\link[stats]{uniroot}}, to find break-even
prices for lifetime licenses.
}
\examples{
# see ?nc_revenue for an example
}
\seealso{
Other wrapper functions for NC results: \code{\link{nc_break_even_yrs}},
  \code{\link{nc_price_lifetime_youth}},
  \code{\link{nc_retain_youth}}, \code{\link{nc_retain}},
  \code{\link{nc_revenue}}
}
\concept{wrapper functions for NC results}
