% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-results.R
\name{nc_break_even_yrs}
\alias{nc_break_even_yrs}
\title{Calculate number of years it takes to break even (if ever)}
\usage{
nc_break_even_yrs(annual_revenue, lifetime_revenue, ignore_wsfr = TRUE)
}
\arguments{
\item{annual_revenue}{target revenue to solve lifetime break-even price}

\item{lifetime_revenue}{revenue for lifetime scenario}

\item{ignore_wsfr}{if TRUE, don't include WSFR dollars in break-even price}
}
\description{
Generally it will take some number of years for cumulative revenue in
the lifetime scenario to catch up with cumulative revenue in the annual
scenario. If the lifetime scenario does catch up, the years to break even
represents the number of years it takes for this to happen.
}
\examples{
# see ?nc_revenue for an example
}
\seealso{
Other wrapper functions for NC results: \code{\link{nc_break_even}},
  \code{\link{nc_price_lifetime_youth}},
  \code{\link{nc_retain}}, \code{\link{nc_revenue}}
}
\concept{wrapper functions for NC results}
