% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-results.R
\name{nc_price_lifetime_youth}
\alias{nc_price_lifetime_youth}
\title{Calculate compounded price for youth lifetimes}
\usage{
nc_price_lifetime_youth(prices, return_life, inflation,
  youth_ages = 0:15)
}
\arguments{
\item{prices}{data frame of lifetime prices by age with at least
2 variables: current_age and price_lifetime}

\item{return_life}{percentage return from lifetime fund}

\item{inflation}{inflation rate for depreciation of lifetime fund}

\item{youth_ages}{if not NULL, assumes for youths that the fund is able to
compound until adulthood (when the agency will begin drawing revenue). See
\code{\link{nc_price_lifetime_youth}} for details.}
}
\description{
This function is based on the assumption that youth lifetime sales are
treated differently than those for adults. In NC, the revenue for a youth sale
cannot be withdrawn (from the fund) until age 16. This implies that a youth
lifetime license sale has greater value (since it is able to compound). Therefore
the effective price is higher than the sticker price for the purpose of
calculating present value of the lifetime license.
}
\examples{
library(dplyr)
library(ggplot2)
prices <- tibble(
    current_age = c(0:15, 16:63),
    price_lifetime = c(rep(250, 64))
)
prices2 <- nc_price_lifetime_youth(prices, 0.05, 0.022)
ggplot(prices, aes(current_age, price_lifetime)) +
    geom_line() +
    geom_line(data = prices2, color = "blue") +
    scale_y_continuous(limits = c(0, 600))
}
\seealso{
Other wrapper functions for NC results: \code{\link{nc_break_even_yrs}},
  \code{\link{nc_break_even}}, \code{\link{nc_retain}},
  \code{\link{nc_revenue}}
}
\concept{wrapper functions for NC results}
