% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-results.R
\name{nc_retain}
\alias{nc_retain}
\alias{nc_retain_all}
\title{Predict retention across a set of ages for NC methodology}
\usage{
nc_retain(retain_all)

nc_retain_all(history_split, ages, use_observed = FALSE)
}
\arguments{
\item{retain_all}{full set of retention curve results produced by
nc_retain_all()}

\item{history_split}{license history list produced by yrs_zero_split()}

\item{ages}{set of ages for which retention curves will be calculated separately}

\item{use_observed}{if TRUE, return observed (instead of predicted) retention
rates}
}
\description{
Convenience functions for the NC analysis, mainly wrappers for
\code{\link{yrs_result_retain}}
}
\section{Functions}{
\itemize{
\item \code{nc_retain_all}: Predict full retention curves
}}

\examples{
library(dplyr)
library(ggplot2)
data(all_sports, lic, sale)

df_split <- all_sports \%>\%
    yrs_lifetime_join(sale, lic, "sportsman") \%>\%
    yrs_zero_split() \%>\%
    yrs_zero_filter(function(x) filter(x, life_group == "sportsman"))

observe_all <- nc_retain_all(df_split, 52:58, use_observed = TRUE)
retain_all <- nc_retain_all(df_split, 52:58)
retain <- nc_retain(retain_all)
retain \%>\%
    ggplot(aes(current_age, yrs)) +
    geom_point() +
    ggtitle("Predicted years of purchases by current age")

# observed vs predicted for specified ages
observe <- filter(observe_all, current_age == 52)
retain <- filter(retain_all, current_age == 52)
yrs_plot(retain) + geom_point(data = observe) +
    ggtitle("Some noisy data points due to sample size")

observe <- filter(observe_all, current_age == 58)
retain <- filter(retain_all, current_age == 58)
yrs_plot(retain) + geom_point(data = observe) +
    ggtitle("Observed rates are used for old enough buyers",
            "Senior lifetime artifacts emerge at age 65")
}
\seealso{
Other wrapper functions for NC results: \code{\link{nc_break_even_yrs}},
  \code{\link{nc_break_even}},
  \code{\link{nc_price_lifetime_youth}},
  \code{\link{nc_retain_youth}}, \code{\link{nc_revenue}}
}
\concept{wrapper functions for NC results}
