% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-results.R
\name{nc_retain_youth}
\alias{nc_retain_youth}
\title{Simulate retention curves for youths}
\usage{
nc_retain_youth(retain_all, youth_ages = 0:15)
}
\arguments{
\item{retain_all}{full set of retention curve results produced by
nc_retain_all()}

\item{youth_ages}{if not NULL, assumes for youths that the fund is able to
compound until adulthood (when the agency will begin drawing revenue). See
\code{\link{nc_price_lifetime_youth}} for details.}
}
\description{
Based on the assumption that youths won't generate revenue until age 16 (default).
Any youth years are to be ignored for revenue purposes (other than wsfr lifetime dollars).
Retention curves are unknown for youth buyers. The simple assumption is that
their curves will match 16-year-olds (at age 16). This likely overestimates
their future participation.
}
\seealso{
Other wrapper functions for NC results: \code{\link{nc_break_even_yrs}},
  \code{\link{nc_break_even}},
  \code{\link{nc_price_lifetime_youth}},
  \code{\link{nc_retain}}, \code{\link{nc_revenue}}
}
\concept{wrapper functions for NC results}
