% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-YrsAnnual.R
\name{yrs_avidity}
\alias{yrs_avidity}
\title{Estimate a simple avidity measure}
\usage{
yrs_avidity(history, num_yrs = 3, drop_na_yrs = TRUE)
}
\arguments{
\item{history}{license history table (1 row per customer per year)}

\item{num_yrs}{number of years to use for defining avidity}

\item{drop_na_yrs}{if TRUE, records for the years in which avidity can't
be calculated will be dropped (e.g., drops first 3 years by default)}
}
\description{
Avidity is based on years of license purchases prior to the current year.
}
\examples{
data(all_sports)
x <- yrs_avidity(all_sports)
summary(x$num_years_held)
}
\seealso{
Other functions to estimate annual license buying: \code{\link{yrs_calc_avg}},
  \code{\link{yrs_calc}}, \code{\link{yrs_fit}},
  \code{\link{yrs_lifetime}}, \code{\link{yrs_plot}},
  \code{\link{yrs_predict_avg}}, \code{\link{yrs_predict}},
  \code{\link{yrs_result}}, \code{\link{yrs_zero}}
}
\concept{functions to estimate annual license buying}
