% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-YrsAnnual.R
\name{yrs_calc}
\alias{yrs_calc}
\alias{yrs_calc_renew}
\alias{yrs_calc_renew_one}
\alias{yrs_calc_retain}
\title{Calculate license buying}
\usage{
yrs_calc_renew(history_split)

yrs_calc_renew_one(history_split, year0)

yrs_calc_retain(history_split, ...)
}
\arguments{
\item{history_split}{license history list produced by yrs_zero_split()}

\item{year0}{year for defining customers for whom retention will be calculated}

\item{...}{optional grouping variables}
}
\description{
Functions to calculate retention for modelling with retain or renew methods.
These should only be run after preparation using \code{\link{yrs_zero_split}}
}
\section{Functions}{
\itemize{
\item \code{yrs_calc_renew}: Calculate license buying - renewal

Appends a new "renew" variable (1=yes, 0=no) based on whether customers from
year zero held licenses in subsequent years. The customers
from year0 will have one record for all subsequent years.

\item \code{yrs_calc_renew_one}: Calculate license buying - renewal year0

Identify renewal for selected year0. Intended to be run from
\code{\link{yrs_calc_renew}}

\item \code{yrs_calc_retain}: Calculate license buying - retention rate

This essentially summarizes a customer's probability of renewing in future
years.
}}

\examples{
library(dplyr)
data(hunt)
hunt_split <- yrs_zero_split(hunt) \%>\%
   yrs_zero_filter(function(x) filter(x, age_year \%in\% 30:50))

yrs_calc_renew_one(hunt_split, 2008)
yrs_calc_renew(hunt_split)
yrs_calc_retain(hunt_split)

library(ggplot2)
retain <- yrs_calc_retain(hunt_split, year0)
ggplot(retain, aes(years_since, pct, color = year0, size = n0)) +
    geom_point() +
    ggtitle("We see some variation in retention curves depending on year zero",
            "more distant future years (years_since) have fewer observations")
}
\seealso{
Other functions to estimate annual license buying: \code{\link{yrs_avidity}},
  \code{\link{yrs_calc_avg}}, \code{\link{yrs_fit}},
  \code{\link{yrs_lifetime}}, \code{\link{yrs_plot}},
  \code{\link{yrs_predict_avg}}, \code{\link{yrs_predict}},
  \code{\link{yrs_result}}, \code{\link{yrs_zero}}
}
\concept{functions to estimate annual license buying}
