% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-YrsAnnual.R
\name{yrs_calc_avg}
\alias{yrs_calc_avg}
\title{Calculate license buying - years by age}
\usage{
yrs_calc_avg(history, ..., yrs = NULL, ages = 16:64,
  age_cohort = function(x) {     cut(x, c(-Inf, 15, 19, 24, 29, 34, 39,
  44, 49, 54, 59, 64, Inf)) })
}
\arguments{
\item{history}{license history table with cust_id, year, age_year}

\item{...}{optional grouping variables}

\item{yrs}{set of years that will be averaged. If NULL, uses the most recent
5 years}

\item{ages}{ages to include, based on first year to be averaged}

\item{age_cohort}{function to estimate age_cohort based on history$birth_year
in the first year to be averaged}
}
\description{
Attempting here to replicate the approach used for the OK lifetime license
study for characterizing particiation over a sportsperson's lifetime. Takes the
average number of years (by cohort) that sportspersons held a license, conditional
on having held a license in at least one of the selected years.
}
\examples{
data(all_sports)
yrs_calc_avg(all_sports)
}
\seealso{
Other functions to estimate annual license buying: \code{\link{yrs_avidity}},
  \code{\link{yrs_calc}}, \code{\link{yrs_fit}},
  \code{\link{yrs_lifetime}}, \code{\link{yrs_plot}},
  \code{\link{yrs_predict_avg}}, \code{\link{yrs_predict}},
  \code{\link{yrs_result}}, \code{\link{yrs_zero}}
}
\concept{functions to estimate annual license buying}
