% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-YrsAnnual.R
\name{yrs_fit}
\alias{yrs_fit}
\alias{yrs_fit_renew}
\alias{yrs_fit_renew_prep}
\alias{yrs_fit_retain}
\alias{yrs_fit_retain_gam}
\title{Fit license buying}
\usage{
yrs_fit_renew(train_df, model = function(train_df) {     glm(renew ~
  num_years_held + log(years_since), family = binomial(link = "logit"),
  data = train_df) })

yrs_fit_renew_prep(train_df)

yrs_fit_retain(train_df, model = function(train_df) lm(pct ~
  log(years_since), data = train_df))

yrs_fit_retain_gam(train_df, model = function(train_df) gam(pct ~
  s(years_since, sp = 0.1), data = train_df))
}
\arguments{
\item{train_df}{data frame to use as training dataset}

\item{model}{function used to model training dataset}
}
\description{
Convenience functions for running a model using selected defaults.
}
\section{Functions}{
\itemize{
\item \code{yrs_fit_renew}: Fit license buying - renewal, logit

\item \code{yrs_fit_renew_prep}: Helper function to prepare data for renewal fit

\item \code{yrs_fit_retain}: Fit license buying - retention, log

\item \code{yrs_fit_retain_gam}: Fit license buying - retention, gam
}}

\examples{
library(dplyr)
data(all_sports)
all_sports <- yrs_avidity(all_sports, drop_na_yrs = FALSE)
df_split <- yrs_zero_split(all_sports) \%>\%
    yrs_zero_filter(function(x) filter(x, age_year == 30))

train_df <- yrs_calc_renew(df_split)
yrs_fit_renew(train_df)

train_df <- yrs_calc_retain(df_split)
yrs_fit_retain(train_df)
yrs_fit_retain_gam(train_df)
}
\seealso{
Other functions to estimate annual license buying: \code{\link{yrs_avidity}},
  \code{\link{yrs_calc_avg}}, \code{\link{yrs_calc}},
  \code{\link{yrs_lifetime}}, \code{\link{yrs_plot}},
  \code{\link{yrs_predict_avg}}, \code{\link{yrs_predict}},
  \code{\link{yrs_result}}, \code{\link{yrs_zero}}
}
\concept{functions to estimate annual license buying}
