% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-YrsAnnual.R
\name{yrs_lifetime}
\alias{yrs_lifetime}
\alias{yrs_lifetime_join}
\alias{yrs_lifetime_split}
\title{Prepare data for comparing lifetime vs annual buyers}
\usage{
yrs_lifetime_join(history, sale, lic, life_slct)

yrs_lifetime_split(history_split)
}
\arguments{
\item{history}{license history data frame}

\item{sale, lic}{license data frames}

\item{life_slct}{name of group to include from lic$life_group (e.g., "sportsman").
Only one value should be specified.}

\item{history_split}{license history list produced by yrs_zero_split()}
}
\description{
For preparing to run \code{\link{yrs_result_renew}} (logistic regression) with
annual vs. lifetime buyers separately.
}
\section{Functions}{
\itemize{
\item \code{yrs_lifetime_join}: Add lifetime sales info for selected license group

\item \code{yrs_lifetime_split}: Split into lifetime vs. non-lifetime buyers
}}

\examples{
\dontrun{
library(DBI)
library(tidyverse)
f <- "E:/SA/Data-production/NCWRC-19-01/license.sqlite3"
con <- dbConnect(RSQLite::SQLite(), f)
all_sports <- tbl(con, "all_sports") \%>\% collect()
lic <- tbl(con, "lic") \%>\% select(lic_id, life_group, duration) \%>\% collect()
sale <- tbl(con, "sale") \%>\% select(lic_id, year, cust_id) \%>\% collect()
dbDisconnect(con)

all_sports <- all_sports \%>\%
    yrs_avidity() \%>\%
    yrs_lifetime_join(sale, lic, "sportsman")

df_split <- all_sports \%>\%
    yrs_zero_split() \%>\%
    yrs_zero_filter(function(x) filter(x, age_year \%in\% 25:35,
                    life_group == "sportsman"))
life_split <- yrs_lifetime_split(df_split)

renew_life <- yrs_result_renew(life_split$annual, life_split$lifetime) \%>\%
    mutate(method = "renew_life")
renew <- yrs_result_renew(life_split$annual)

bind_rows(renew, renew_life) \%>\% yrs_plot() +
   ggtitle("Sportsman annual vs lifetime estimated years")
}
}
\seealso{
Other functions to estimate annual license buying: \code{\link{yrs_avidity}},
  \code{\link{yrs_calc_avg}}, \code{\link{yrs_calc}},
  \code{\link{yrs_fit}}, \code{\link{yrs_plot}},
  \code{\link{yrs_predict_avg}}, \code{\link{yrs_predict}},
  \code{\link{yrs_result}}, \code{\link{yrs_zero}}
}
\concept{functions to estimate annual license buying}
