% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-YrsAnnual.R
\name{yrs_plot}
\alias{yrs_plot}
\title{Plot results for predicting license buying}
\usage{
yrs_plot(result_df, include_label = TRUE, xlim = 74, grp = method)
}
\arguments{
\item{result_df}{results returned from \code{\link{yrs_result}}}

\item{include_label}{if TRUE, displays the method name and years estimate
next to the line}

\item{xlim}{maximum x-axis limit for plot}

\item{grp}{unquoted name of grouping variable}
}
\description{
It's useful to visualize the results with retention estimates over time,
especially for comparing alternative methods. This function takes the result
of \code{\link{yrs_result}} and plots with some useful labelling and scaling.
}
\examples{
library(dplyr)
library(ggplot2)

data(all_sports)
avg <- yrs_result_avg(all_sports)

df_split <- all_sports \%>\%
    yrs_avidity(drop_na_yrs = FALSE) \%>\%
    yrs_zero_split() \%>\%
    yrs_zero_filter(function(x) filter(x, age_year \%in\% 25:35))

# comparing retention-based vs avg-years-based methods
observe <- yrs_result_observe(df_split)
retain <- yrs_result_retain(df_split)
max <- yrs_result_max(df_split)
bind_rows(retain, max, avg) \%>\% yrs_plot() + geom_point(data = observe)

# comparing retain & renew methods
renew <- yrs_result_renew(df_split)
bind_rows(retain, renew) \%>\% yrs_plot() + geom_point(data = observe)

# using a GAM for retain
retain <- df_split \%>\%
    yrs_result_retain(model = function(x) yrs_fit_retain_gam(x))
bind_rows(retain, max, avg) \%>\% yrs_plot() + geom_point(data = observe)
}
\seealso{
Other functions to estimate annual license buying: \code{\link{yrs_avidity}},
  \code{\link{yrs_calc_avg}}, \code{\link{yrs_calc}},
  \code{\link{yrs_fit}}, \code{\link{yrs_lifetime}},
  \code{\link{yrs_predict_avg}}, \code{\link{yrs_predict}},
  \code{\link{yrs_result}}, \code{\link{yrs_zero}}
}
\concept{functions to estimate annual license buying}
