% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-YrsAnnual.R
\name{yrs_predict}
\alias{yrs_predict}
\alias{yrs_predict_retain}
\alias{yrs_predict_renew}
\title{Predict license buying}
\usage{
yrs_predict_retain(predict_df, model_fit, method = "retain")

yrs_predict_renew(predict_df, model_fit, method = "renew")
}
\arguments{
\item{predict_df}{data frame to use for prediction}

\item{model_fit}{model to run \code{\link[stats]{predict}}}

\item{method}{character name of method used (for plotting)}
}
\description{
Convenience functions for predicting on a dataset using a specified
model fit.
}
\section{Functions}{
\itemize{
\item \code{yrs_predict_retain}: Predict license buying - retention

\item \code{yrs_predict_renew}: Predict license buying - renewal
}}

\examples{
library(dplyr)
library(ggplot2)
data(all_sports)

df_split <- all_sports \%>\%
    yrs_avidity() \%>\%
    yrs_zero_split() \%>\%
    yrs_zero_filter(function(x) filter(x, age_year \%in\% 30:50))

# log trend fit of retention rates
train_df <- yrs_calc_retain(df_split)
model_fit <- yrs_fit_retain(train_df)
out <- yrs_predict_retain(data.frame(years_since = 1:40), model_fit)
sum(out$pct) # estimated annual license purchases over 40 years
ggplot(out, aes(years_since, pct)) + geom_line() + geom_point(data = train_df)

# logistic regression based on individual-level renewal
train_df2 <- yrs_calc_renew(df_split)
model_fit2 <- yrs_fit_renew(train_df2)
predict_df2 <- filter(df_split$year0, year == 2011) \%>\% select(num_years_held)
out2 <- 1:40 \%>\%
    sapply(function(i) mutate(predict_df2, years_since = i), simplify = FALSE) \%>\%
    bind_rows() \%>\%
    yrs_predict_renew(model_fit2)
sum(out2$pct)
ggplot(out2, aes(years_since, pct)) + geom_line() + geom_point(data = train_df)
}
\seealso{
Other functions to estimate annual license buying: \code{\link{yrs_avidity}},
  \code{\link{yrs_calc_avg}}, \code{\link{yrs_calc}},
  \code{\link{yrs_fit}}, \code{\link{yrs_lifetime}},
  \code{\link{yrs_plot}}, \code{\link{yrs_predict_avg}},
  \code{\link{yrs_result}}, \code{\link{yrs_zero}}
}
\concept{functions to estimate annual license buying}
