% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{cust}
\alias{cust}
\title{Data: customer sample}
\format{data frame}
\description{
Data: customer sample
}
\seealso{
Other Data: \code{\link{all_sports}}, \code{\link{lic}},
  \code{\link{retain_all}}, \code{\link{retain}},
  \code{\link{sale}}
}
\concept{Data}
\keyword{datasets}
