% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetime.R
\docType{data}
\name{lic}
\alias{lic}
\title{Data: license types of interest for analysis}
\format{data frame}
\description{
Data: license types of interest for analysis
}
\seealso{
Other Data: \code{\link{cust}}, \code{\link{hunt}},
  \code{\link{retain}}, \code{\link{sale}}
}
\concept{Data}
\keyword{datasets}
