% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-results.R
\name{nc_break_even}
\alias{nc_break_even}
\title{Find lifetime price (by current_age) where Revenue|A == Revenue|L}
\usage{
nc_break_even(annual_revenue, wsfr_amount, min_amount, return_life,
  inflation, perpetuity = TRUE, ignore_wsfr = TRUE,
  youth_ages = 0:15)
}
\arguments{
\item{annual_revenue}{target revenue to solve lifetime break-even price}

\item{wsfr_amount}{numeric estimated amount for aid dollars (use SFRF for
fishing, WRF for hunting)}

\item{min_amount}{numeric the minimum expenditure that will count for a
certified hunter/angler (use 2 for hunt/fish and 4 for combo)}

\item{return_life}{percentage return from lifetime fund}

\item{inflation}{inflation rate for depreciation of lifetime fund}

\item{perpetuity}{if TRUE, use a perpetuity calculation
(\url{https://en.wikipedia.org/wiki/Perpetuity}) instead of
\code{\link{present_value}}}

\item{ignore_wsfr}{if TRUE, don't include WSFR dollars in break-even price}

\item{youth_ages}{if not NULL, assumes for youths that the fund is able to
compound until adulthood (when the agency will begin drawing revenue). See
\code{\link{nc_price_lifetime_youth}} for details.}
}
\description{
This uses a solver function, \code{\link[stats]{uniroot}}, to find break-even
prices for lifetime licenses given predicted annual revenue.
}
\examples{
# see ?nc_revenue for an example
}
\seealso{
Other wrapper functions for NC results: \code{\link{nc_break_even_yrs}},
  \code{\link{nc_price_lifetime_youth}},
  \code{\link{nc_retain}}, \code{\link{nc_revenue}}
}
\concept{wrapper functions for NC results}
