% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-results.R
\name{nc_retain}
\alias{nc_retain}
\alias{nc_retain_all}
\title{Predict retention across a set of ages for NC methodology}
\usage{
nc_retain(retain_all)

nc_retain_all(history_split, ages)
}
\arguments{
\item{retain_all}{full set of retention curve results produced by 
nc_retain_all()}

\item{history_split}{license history list produced by yrs_zero_split()}

\item{ages}{set of ages for which retention curves will be calculated separately}
}
\description{
Convenience functions for the NC analysis, mainly wrappers for 
\code{\link{yrs_result_retain}}
}
\section{Functions}{
\itemize{
\item \code{nc_retain_all}: Predict full retention curves
}}

\examples{
library(dplyr)
library(ggplot2)
data(hunt, lic, sale)

hunt_split <- hunt \%>\%
    yrs_lifetime_join(sale, lic, "sportsman") \%>\%
    yrs_zero_split() \%>\%
    yrs_zero_filter(function(x) filter(x, life_group == "sportsman"))
    
retain <- nc_retain_all(hunt_split, 40:50) \%>\% nc_retain()
retain \%>\%
    ggplot(aes(current_age, yrs)) +
    geom_point() +
    ggtitle("Predicted years of purchases by current age")
}
\seealso{
Other wrapper functions for NC results: \code{\link{nc_retain_youth}},
  \code{\link{nc_revenue}}
}
\concept{wrapper functions for NC results}
