% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-results.R
\name{nc_retain_youth}
\alias{nc_retain_youth}
\title{Get expected annual revenue-equivalent years for youths}
\usage{
nc_retain_youth(retain, rate = 0.03, ages = 0:15)
}
\arguments{
\item{retain}{retention results like those produced by \code{\link{nc_retain}}}

\item{rate}{depreciation rate}

\item{ages}{ages for youth (i.e., those that haven't been already estimated)}
}
\description{
This is applied as a way to include annual revenue estimates for infants and
youths. It essentially applies a depreciation to the annual revenue estimates,
providing break-even estimates that are lower for youths.
}
\seealso{
Other wrapper functions for NC results: \code{\link{nc_retain}},
  \code{\link{nc_revenue}}
}
\concept{wrapper functions for NC results}
