% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-results.R
\name{nc_revenue}
\alias{nc_revenue}
\alias{nc_revenue_annual}
\alias{nc_revenue_annual2}
\alias{nc_retain_youth}
\alias{nc_revenue_lifetime}
\title{Calculate revenue for annual vs lifetime scenarios}
\usage{
nc_revenue_annual(retain, prices, wsfr_amount, min_amount, senior_price,
  senior_age = 65, age_cutoff = 80, youth_ages = 0:15)

nc_revenue_annual2(retain_all, prices, wsfr_amount, min_amount,
  senior_price, senior_age = 65, age_cutoff = 80, youth_ages = 0:15)

nc_retain_youth(retain_all, youth_ages = 0:15)

nc_revenue_lifetime(prices, wsfr_amount, min_amount, return_life,
  inflation, perpetuity = TRUE, youth_ages = 0:15)
}
\arguments{
\item{retain}{data frame of predicted years by age like that produced by
\code{\link{nc_retain}}}

\item{prices}{data frame of lifetime prices by age with at least
2 variables: current_age and price_lifetime}

\item{wsfr_amount}{numeric estimated amount for aid dollars (use SFRF for
fishing, WRF for hunting)}

\item{min_amount}{numeric the minimum expenditure that will count for a
certified hunter/angler (use 2 for hunt/fish and 4 for combo)}

\item{senior_price}{numeric price for a senior lifetime license}

\item{senior_age}{numeric age when a participant will be expected to buy
a cheap lifetime license}

\item{age_cutoff}{numeric final age that can be counted for lifetime
license-based WSFR dollars}

\item{youth_ages}{if not NULL, assumes for youths that the fund is able to
compound until adulthood (when the agency will begin drawing revenue). See
\code{\link{nc_price_lifetime_youth}} for details.}

\item{retain_all}{data frame of predicted years by age like that produced by
\code{\link{nc_retain_all}}}

\item{return_life}{percentage return from lifetime fund}

\item{inflation}{inflation rate for depreciation of lifetime fund}

\item{perpetuity}{if TRUE, use a perpetuity calculation
(\url{https://en.wikipedia.org/wiki/Perpetuity}) instead of
\code{\link{present_value}}}
}
\description{
Calculate revenue for annual vs lifetime scenarios
}
\section{Functions}{
\itemize{
\item \code{nc_revenue_annual}: Revenue for annual scenario (R|A)

\item \code{nc_revenue_annual2}: Revenue for annual scenario (Revenue|A)

\item \code{nc_retain_youth}: Simulate retention curves for youths

\item \code{nc_revenue_lifetime}: Revenue for lifetime scenario (R|L)
}}

\examples{
library(dplyr)
library(ggplot2)
data(retain)

# parameters
price_annual <- 36
price_lifetime <- 250
senior_price <- 15 # lifetime license price at age 65
wsfr_amount <- 16.65 # WSFR estimated aid amount per hunter
min_amount <- 2 # WSFR minimum revenue for certified hunter calculations
return_life <- 0.05 # annual return to lifetime fund
inflation <- 0.02 # inflation rate for fund depreciation
prices <- tibble(
    current_age = 0:63, price_lifetime = rep(price_lifetime, 64),
    price_annual = rep(price_annual, 64)
)

# annual
annual <- retain \%>\%
    nc_revenue_annual(prices, wsfr_amount, min_amount, senior_price)
ggplot(annual, aes(current_age, revenue_annual, fill = stream)) + geom_col()

# lifetime
lifetime <- prices \%>\%
    nc_revenue_lifetime(wsfr_amount, min_amount, return_life, inflation)
ggplot(lifetime, aes(current_age, revenue_lifetime, fill = stream)) +
    geom_col()

# compare
left_join(lifetime, annual, by = c("current_age", "stream")) \%>\%
    mutate(revenue_effect = revenue_lifetime - revenue_annual) \%>\%
    ggplot(aes(current_age, revenue_effect, fill = stream)) +
    geom_col()

# break-even
x <- nc_break_even(annual, wsfr_amount, min_amount, return_life, inflation)
ggplot(x, aes(current_age, break_even)) + geom_line()

x <- nc_break_even(annual, wsfr_amount, min_amount, return_life, inflation,
                   ignore_wsfr = FALSE)
ggplot(x, aes(current_age, break_even)) + geom_line()

# years to break-even

}
\seealso{
Other wrapper functions for NC results: \code{\link{nc_break_even_yrs}},
  \code{\link{nc_break_even}},
  \code{\link{nc_price_lifetime_youth}},
  \code{\link{nc_retain}}
}
\concept{wrapper functions for NC results}
