% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{retain_all}
\alias{retain_all}
\title{Data: Sportsman Retention years results by current & future age}
\format{data frame}
\description{
Data: Sportsman Retention years results by current & future age
}
\seealso{
Other Data: \code{\link{all_sports}}, \code{\link{cust}},
  \code{\link{lic}}, \code{\link{retain}},
  \code{\link{sale}}
}
\concept{Data}
\keyword{datasets}
