% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{sale}
\alias{sale}
\title{Data: hunting sales for license types of interest in sample}
\format{data frame}
\description{
Data: hunting sales for license types of interest in sample
}
\seealso{
Other Data: \code{\link{cust}}, \code{\link{hunt}},
  \code{\link{lic}}, \code{\link{retain_all}},
  \code{\link{retain}}
}
\concept{Data}
\keyword{datasets}
